// src/generated/stoporders.ts
export enum StopOrderDirection {
  /** STOP_ORDER_DIRECTION_UNSPECIFIED - Значение не указано */
  STOP_ORDER_DIRECTION_UNSPECIFIED = 0,
  /** STOP_ORDER_DIRECTION_BUY - Покупка */
  STOP_ORDER_DIRECTION_BUY = 1,
  /** STOP_ORDER_DIRECTION_SELL - Продажа */
  STOP_ORDER_DIRECTION_SELL = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum StopOrderExpirationType {
  /** STOP_ORDER_EXPIRATION_TYPE_UNSPECIFIED - Значение не указано */
  STOP_ORDER_EXPIRATION_TYPE_UNSPECIFIED = 0,
  /** STOP_ORDER_EXPIRATION_TYPE_GOOD_TILL_CANCEL - Действительно до отмены */
  STOP_ORDER_EXPIRATION_TYPE_GOOD_TILL_CANCEL = 1,
  /** STOP_ORDER_EXPIRATION_TYPE_GOOD_TILL_DATE - Действительно до даты снятия */
  STOP_ORDER_EXPIRATION_TYPE_GOOD_TILL_DATE = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum StopOrderType {
  /** STOP_ORDER_TYPE_UNSPECIFIED - Значение не указано */
  STOP_ORDER_TYPE_UNSPECIFIED = 0,
  /** STOP_ORDER_TYPE_TAKE_PROFIT - Take-profit заявка */
  STOP_ORDER_TYPE_TAKE_PROFIT = 1,
  /** STOP_ORDER_TYPE_STOP_LOSS - Stop-loss заявка */
  STOP_ORDER_TYPE_STOP_LOSS = 2,
  /** STOP_ORDER_TYPE_STOP_LIMIT - Stop-limit заявка */
  STOP_ORDER_TYPE_STOP_LIMIT = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum ExchangeOrderType {
  /** EXCHANGE_ORDER_TYPE_UNSPECIFIED - Значение не указано */
  EXCHANGE_ORDER_TYPE_UNSPECIFIED = 0,
  /** EXCHANGE_ORDER_TYPE_MARKET - Заявка по рыночной цене */
  EXCHANGE_ORDER_TYPE_MARKET = 1,
  /** EXCHANGE_ORDER_TYPE_LIMIT - Лимитная заявка */
  EXCHANGE_ORDER_TYPE_LIMIT = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum TakeProfitType {
  /** TAKE_PROFIT_TYPE_UNSPECIFIED - Значение не указано */
  TAKE_PROFIT_TYPE_UNSPECIFIED = 0,
  /** TAKE_PROFIT_TYPE_REGULAR - Обычная заявка (значение по умолчанию) */
  TAKE_PROFIT_TYPE_REGULAR = 1,
  /** TAKE_PROFIT_TYPE_TRAILING - Трейлинг-стоп */
  TAKE_PROFIT_TYPE_TRAILING = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum TrailingValueType {
  /** TRAILING_VALUE_UNSPECIFIED - Значение не указано */
  TRAILING_VALUE_UNSPECIFIED = 0,
  /** TRAILING_VALUE_ABSOLUTE - Абсолютное значение в единицах цены */
  TRAILING_VALUE_ABSOLUTE = 1,
  /** TRAILING_VALUE_RELATIVE - Относительное значение в процентах */
  TRAILING_VALUE_RELATIVE = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface PostStopOrderRequest {
  /**
   * Deprecated Figi-идентификатор инструмента. Необходимо использовать instrument_id.
   *
   * @deprecated
   */
  figi?:
    | string
    | undefined;
  /** Количество лотов */
  quantity: number;
  /** Цена за 1 инструмент. Для получения стоимости лота требуется умножить на лотность инструмента. */
  price?:
    | Quotation
    | undefined;
  /** Стоп-цена заявки за 1 инструмент. Для получения стоимости лота требуется умножить на лотность инструмента. */
  stopPrice?:
    | Quotation
    | undefined;
  /** Направление операции */
  direction: StopOrderDirection;
  /** Номер счёта */
  accountId: string;
  /** Тип экспирации заявки */
  expirationType: StopOrderExpirationType;
  /** Тип заявки */
  stopOrderType: StopOrderType;
  /** Дата и время окончания действия стоп-заявки в часовом поясе UTC. **Для ExpirationType = GoodTillDate заполнение обязательно, для GoodTillCancel игнорируется**. */
  expireDate?:
    | Date
    | undefined;
  /** Идентификатор инструмента, принимает значения Figi или instrument_uid. */
  instrumentId: string;
  /** Тип дочерней биржевой заявки для тейкпрофита */
  exchangeOrderType: ExchangeOrderType;
  /** Подтип стоп-заявки TakeProfit */
  takeProfitType: TakeProfitType;
  /** Массив с параметрами трейлинг-стопа */
  trailingData:
    | PostStopOrderRequest_TrailingData
    | undefined;
  /** Тип цены */
  priceType: PriceType;
  /** Идентификатор запроса выставления поручения для целей идемпотентности в формате UID. Максимальная длина 36 символов. */
  orderId: string;
}

export interface PostStopOrderRequest_TrailingData {
  /** Отступ */
  indent:
    | Quotation
    | undefined;
  /** Тип величины отступа */
  indentType: TrailingValueType;
  /** Размер защитного спреда */
  spread:
    | Quotation
    | undefined;
  /** Тип величины защитного спреда */
  spreadType: TrailingValueType;
}

/** Результат выставления стоп-заявки. */
export interface PostStopOrderResponse {
  /** Уникальный идентификатор стоп-заявки */
  stopOrderId: string;
  /** Идентификатор ключа идемпотентности, переданный клиентом, в формате UID. Максимальная длина 36 символов. */
  orderRequestId: string;
  /** Метадата */
  responseMetadata: ResponseMetadata | undefined;
}
// ...
export const PostStopOrderRequest_TrailingData = {
  encode(message: PostStopOrderRequest_TrailingData, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.indent !== undefined) {
      Quotation.encode(message.indent, writer.uint32(10).fork()).ldelim();
    }
    if (message.indentType !== 0) {
      writer.uint32(16).int32(message.indentType);
    }
    if (message.spread !== undefined) {
      Quotation.encode(message.spread, writer.uint32(26).fork()).ldelim();
    }
    if (message.spreadType !== 0) {
      writer.uint32(32).int32(message.spreadType);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PostStopOrderRequest_TrailingData {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePostStopOrderRequest_TrailingData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.indent = Quotation.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.indentType = reader.int32() as any;
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.spread = Quotation.decode(reader, reader.uint32());
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.spreadType = reader.int32() as any;
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PostStopOrderRequest_TrailingData {
    return {
      indent: isSet(object.indent) ? Quotation.fromJSON(object.indent) : undefined,
      indentType: isSet(object.indentType) ? trailingValueTypeFromJSON(object.indentType) : 0,
      spread: isSet(object.spread) ? Quotation.fromJSON(object.spread) : undefined,
      spreadType: isSet(object.spreadType) ? trailingValueTypeFromJSON(object.spreadType) : 0,
    };
  },

  toJSON(message: PostStopOrderRequest_TrailingData): unknown {
    const obj: any = {};
    if (message.indent !== undefined) {
      obj.indent = Quotation.toJSON(message.indent);
    }
    if (message.indentType !== 0) {
      obj.indentType = trailingValueTypeToJSON(message.indentType);
    }
    if (message.spread !== undefined) {
      obj.spread = Quotation.toJSON(message.spread);
    }
    if (message.spreadType !== 0) {
      obj.spreadType = trailingValueTypeToJSON(message.spreadType);
    }
    return obj;
  },

  create(base?: DeepPartial<PostStopOrderRequest_TrailingData>): PostStopOrderRequest_TrailingData {
    return PostStopOrderRequest_TrailingData.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PostStopOrderRequest_TrailingData>): PostStopOrderRequest_TrailingData {
    const message = createBasePostStopOrderRequest_TrailingData();
    message.indent = (object.indent !== undefined && object.indent !== null)
      ? Quotation.fromPartial(object.indent)
      : undefined;
    message.indentType = object.indentType ?? 0;
    message.spread = (object.spread !== undefined && object.spread !== null)
      ? Quotation.fromPartial(object.spread)
      : undefined;
    message.spreadType = object.spreadType ?? 0;
    return message;
  },
};

function createBasePostStopOrderResponse(): PostStopOrderResponse {
  return { stopOrderId: "", orderRequestId: "", responseMetadata: undefined };
}

export const PostStopOrderResponse = {
  encode(message: PostStopOrderResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.stopOrderId !== "") {
      writer.uint32(10).string(message.stopOrderId);
    }
    if (message.orderRequestId !== "") {
      writer.uint32(18).string(message.orderRequestId);
    }
    if (message.responseMetadata !== undefined) {
      ResponseMetadata.encode(message.responseMetadata, writer.uint32(2034).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PostStopOrderResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePostStopOrderResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.stopOrderId = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.orderRequestId = reader.string();
          continue;
        case 254:
          if (tag !== 2034) {
            break;
          }

          message.responseMetadata = ResponseMetadata.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PostStopOrderResponse {
    return {
      stopOrderId: isSet(object.stopOrderId) ? globalThis.String(object.stopOrderId) : "",
      orderRequestId: isSet(object.orderRequestId) ? globalThis.String(object.orderRequestId) : "",
      responseMetadata: isSet(object.responseMetadata) ? ResponseMetadata.fromJSON(object.responseMetadata) : undefined,
    };
  },

  toJSON(message: PostStopOrderResponse): unknown {
    const obj: any = {};
    if (message.stopOrderId !== "") {
      obj.stopOrderId = message.stopOrderId;
    }
    if (message.orderRequestId !== "") {
      obj.orderRequestId = message.orderRequestId;
    }
    if (message.responseMetadata !== undefined) {
      obj.responseMetadata = ResponseMetadata.toJSON(message.responseMetadata);
    }
    return obj;
  },

  create(base?: DeepPartial<PostStopOrderResponse>): PostStopOrderResponse {
    return PostStopOrderResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PostStopOrderResponse>): PostStopOrderResponse {
    const message = createBasePostStopOrderResponse();
    message.stopOrderId = object.stopOrderId ?? "";
    message.orderRequestId = object.orderRequestId ?? "";
    message.responseMetadata = (object.responseMetadata !== undefined && object.responseMetadata !== null)
      ? ResponseMetadata.fromPartial(object.responseMetadata)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;